사용자 일치 여부 확인

+++ src/main/java/com/sparta/springtask1/service/CommentService.java
@@ -19,28 +19,48 @@ public class CommentService {
     private final CommentRepository commentRepository;
     private final ScheduleService scheduleService;
 
+    // 댓글 생성
     public List<CommentResponseDto> createComment(Long scheduleId, CommentRequestDto requestDto) {
+        // 해당 일정이 있는지 확인
         Schedule schedule = scheduleService.findSchedule(scheduleId);
+        // 댓글을 담을 List 생성
         List<Comment> commentList = new ArrayList<>();
 
+        // 댓글에 아무 내용이 없으면 예외로 던짐
+        if(requestDto.getDetail().isBlank()) {
+            throw new IllegalArgumentException("내용을 입력해주세요");
+        }
+
+        // 댓글 저장
         Comment comment = new Comment(schedule, requestDto);
+
         commentList.add(comment);
 
         commentRepository.saveAll(commentList);
 
+        // 사용자에게 보여지는 댓글 목록 List 생성
         List<CommentResponseDto> responseDtoList = new ArrayList<>();
 
         for(Comment i : commentList) {
             responseDtoList.add(new CommentResponseDto(i));
         }
 
+        // 사용자에게 보여지는 댓글 목록 List 반환
         return responseDtoList;
     }
 
     @Transactional
     public Long updateComment(Long scheduleId, Long id, CommentRequestDto requestDto) {
 
+        // 해당 댓글이 있는지 확인
         Comment comment = findComment(id);
+
+        // 해당 댓글의 아이디가 수정자가 입력한 아이디와 같은지 확인
+        if(!requestDto.getUserId().equals(comment.getUserId())) {
+            throw new IllegalArgumentException("사용자의 아이디가 일치하지 않습니다.");
+        }
+
+        // 댓글 수정
         comment.update(requestDto);
 
         return id;
@@ -48,7 +68,15 @@ public class CommentService {
 
     public Long deleteComment(Long scheduleId, Long id, CommentRequestDto requestDto) {
 
+        // 해당 댓글이 있는지 확인
         Comment comment = findComment(id);
+
+        // 해당 댓글의 아이디가 삭제하려는 사람이 입력한 아이디와 같은지 확인
+        if(!requestDto.getUserId().equals(comment.getUserId())) {
+            throw new IllegalArgumentException("사용자의 아이디가 일치하지 않습니다.");
+        }
+
+        // 댓글 삭제
         commentRepository.delete(comment);
 
         return id;
